requirements.txt
pymupdf
python-dotenv
requests
playwright
azure-ai-documentintelligence
azure-cosmos
azure-storage-blob
